use simple SIR model as test
