+++ src/covid19model/models/base.py
